requirements.txt
streamlit>=1.47.0
requests>=2.32.0
python-dotenv>=1.1.0
python-docx>=0.8.11
openai>=1.0.0 